+++ src/main/java/ddog/daengleserver/application/UserAccountService.java
@@ -3,9 +3,9 @@ package ddog.daengleserver.application;
 import ddog.daengleserver.application.repository.AccountRepository;
 import ddog.daengleserver.application.repository.PetRepository;
 import ddog.daengleserver.application.repository.UserRepository;
-import ddog.daengleserver.domain.Account;
-import ddog.daengleserver.domain.Pet;
-import ddog.daengleserver.domain.User;
+import ddog.daengleserver.domain.account.Account;
+import ddog.daengleserver.domain.account.Pet;
+import ddog.daengleserver.domain.account.User;
 import ddog.daengleserver.presentation.account.dto.request.JoinUserWithPet;
 import ddog.daengleserver.presentation.account.dto.request.JoinUserWithoutPet;
 import ddog.daengleserver.presentation.account.dto.request.UserProfileModifyReq;
