+++ src/main/java/ddog/daengleserver/application/UserService.java
@@ -3,8 +3,8 @@ package ddog.daengleserver.application;
 import ddog.daengleserver.application.repository.CareEstimateRepository;
 import ddog.daengleserver.application.repository.GroomingEstimateRepository;
 import ddog.daengleserver.application.repository.UserRepository;
-import ddog.daengleserver.domain.Pet;
-import ddog.daengleserver.domain.User;
+import ddog.daengleserver.domain.account.Pet;
+import ddog.daengleserver.domain.account.User;
 import ddog.daengleserver.domain.estimate.CareEstimate;
 import ddog.daengleserver.domain.estimate.GroomingEstimate;
 import ddog.daengleserver.presentation.estimate.dto.response.CareEstimateDetails;
