+++ src/main/java/ddog/daengleserver/application/repository/AccountRepository.java
@@ -1,7 +1,7 @@
 package ddog.daengleserver.application.repository;
 
-import ddog.daengleserver.domain.Account;
-import ddog.daengleserver.domain.Role;
+import ddog.daengleserver.domain.account.Account;
+import ddog.daengleserver.domain.account.enums.Role;
 
 public interface AccountRepository {
