+++ src/main/java/ddog/daengleserver/application/repository/PetRepository.java
@@ -1,6 +1,6 @@
 package ddog.daengleserver.application.repository;
 
-import ddog.daengleserver.domain.Pet;
+import ddog.daengleserver.domain.account.Pet;
 
 import java.util.List;
