--- src/main/java/ddog/daengleserver/domain/Gender.java
@@ -1,5 +1,0 @@
-package ddog.daengleserver.domain;
-
-public enum Gender {
-    MALE, FEMALE
-}
