--- src/main/java/ddog/daengleserver/domain/Provider.java
@@ -1,5 +1,0 @@
-package ddog.daengleserver.domain;
-
-public enum Provider {
-    KAKAO
-}
