+++ src/main/java/ddog/daengleserver/domain/account/Vet.java
@@ -1,4 +1,4 @@
-package ddog.daengleserver.domain;
+package ddog.daengleserver.domain.account;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
