+++ src/main/java/ddog/daengleserver/domain/account/enums/Gender.java
@@ -1,0 +1,5 @@
+package ddog.daengleserver.domain.account.enums;
+
+public enum Gender {
+    MALE, FEMALE
+}
