+++ src/main/java/ddog/daengleserver/domain/account/enums/Provider.java
@@ -1,0 +1,5 @@
+package ddog.daengleserver.domain.account.enums;
+
+public enum Provider {
+    KAKAO
+}
