+++ src/main/java/ddog/daengleserver/global/auth/dto/TokenAccountInfoDto.java
@@ -1,7 +1,7 @@
 package ddog.daengleserver.global.auth.dto;
 
-import ddog.daengleserver.domain.Provider;
-import ddog.daengleserver.domain.Role;
+import ddog.daengleserver.domain.account.enums.Provider;
+import ddog.daengleserver.domain.account.enums.Role;
 import ddog.daengleserver.infrastructure.po.AccountJpaEntity;
 import lombok.Builder;
 import lombok.Getter;
