+++ src/main/java/ddog/daengleserver/infrastructure/AccountJpaRepository.java
@@ -1,7 +1,7 @@
 package ddog.daengleserver.infrastructure;
 
-import ddog.daengleserver.domain.Provider;
-import ddog.daengleserver.domain.Role;
+import ddog.daengleserver.domain.account.enums.Provider;
+import ddog.daengleserver.domain.account.enums.Role;
 import ddog.daengleserver.infrastructure.po.AccountJpaEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
