+++ src/main/java/ddog/daengleserver/infrastructure/AccountRepositoryImpl.java
@@ -1,8 +1,8 @@
 package ddog.daengleserver.infrastructure;
 
 import ddog.daengleserver.application.repository.AccountRepository;
-import ddog.daengleserver.domain.Account;
-import ddog.daengleserver.domain.Role;
+import ddog.daengleserver.domain.account.Account;
+import ddog.daengleserver.domain.account.enums.Role;
 import ddog.daengleserver.implementation.AccountException;
 import ddog.daengleserver.implementation.enums.AccountExceptionType;
 import ddog.daengleserver.infrastructure.po.AccountJpaEntity;
