+++ src/main/java/ddog/daengleserver/infrastructure/po/AccountJpaEntity.java
@@ -1,8 +1,8 @@
 package ddog.daengleserver.infrastructure.po;
 
-import ddog.daengleserver.domain.Account;
-import ddog.daengleserver.domain.Provider;
-import ddog.daengleserver.domain.Role;
+import ddog.daengleserver.domain.account.Account;
+import ddog.daengleserver.domain.account.enums.Provider;
+import ddog.daengleserver.domain.account.enums.Role;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
