+++ src/main/java/ddog/daengleserver/presentation/account/dto/request/JoinUserWithoutPet.java
@@ -1,6 +1,6 @@
 package ddog.daengleserver.presentation.account.dto.request;
 
-import ddog.daengleserver.domain.Role;
+import ddog.daengleserver.domain.account.enums.Role;
 import lombok.Builder;
 import lombok.Getter;
