+++ src/main/java/ddog/daengleserver/presentation/estimate/dto/response/UserAndPetsInfo.java
@@ -1,6 +1,6 @@
 package ddog.daengleserver.presentation.estimate.dto.response;
 
-import ddog.daengleserver.domain.Weight;
+import ddog.daengleserver.domain.account.enums.Weight;
 import lombok.Builder;
 import lombok.Getter;
